+++ common/model/src/main/java/is/codion/common/model/table/TableSelectionModel.java
@@ -161,6 +161,11 @@ public interface TableSelectionModel<R> {
 	interface SelectedItems<R> extends Observable<List<R>> {
 
 		/**
+		 * @param indexes the items to select
+		 */
+		void set(Collection<R> items);
+
+		/**
 		 * Sets the items passing the predicate test as the selection
 		 * @param predicate the predicate
 		 */