+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/DefaultFilterTableSelectionModel.java
@@ -442,6 +442,11 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 		}
 
 		@Override
+		public void set(Collection<R> items) {
+			set(new ArrayList<>(requireNonNull(items)));
+		}
+
+		@Override
 		public void set(List<R> items) {
 			if (!isSelectionEmpty()) {
 				clearSelection();